Revert "Ignore hidden files during serialization (#356)"
This reverts commit 70795644142d245360dbe1d4e3b27ea9a76e6493.
