+++ truss/tests/conftest.py
@@ -677,19 +677,6 @@ def custom_model_truss_dir_with_pre_and_post_description(tmp_path):
 
 
 @pytest.fixture
-def custom_model_truss_dir_with_hidden_files(tmp_path):
-    truss_dir_path: Path = tmp_path / "custom_model_truss_dir_with_hidden_files"
-    _ = init(str(truss_dir_path))
-    (truss_dir_path / "__pycache__").mkdir(parents=True, exist_ok=True)
-    (truss_dir_path / ".git").mkdir(parents=True, exist_ok=True)
-    (truss_dir_path / "__pycache__" / "test.cpython-38.pyc").touch()
-    (truss_dir_path / ".DS_Store").touch()
-    (truss_dir_path / ".git" / ".test_file").touch()
-    (truss_dir_path / "data" / "test_file").write_text("123456789")
-    yield truss_dir_path
-
-
-@pytest.fixture
 def custom_model_truss_dir_for_gpu(tmp_path):
     dir_path = tmp_path / "custom_truss"
     handle = init(str(dir_path))
