+++ truss/truss_handle.py
@@ -44,7 +44,7 @@ from truss.contexts.image_builder.training_image_builder import (
 from truss.contexts.local_loader.load_model_local import LoadModelLocal
 from truss.contexts.local_loader.local_server_loader import LocalServerLoader
 from truss.contexts.local_loader.train_local import LocalTrainer
-from truss.decorators import proxy_to_shadow
+from truss.decorators import proxy_to_shadow_if_scattered
 from truss.docker import (
     Docker,
     DockerStates,
@@ -120,7 +120,7 @@ class TrussHandle:
 
         return requests.post(url, json=request)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def build_docker_build_context(self, build_dir: Optional[Path] = None):
         build_dir_path = Path(build_dir) if build_dir is not None else None
         image_builder = ServingImageBuilderContext.run(self._truss_dir)
@@ -130,7 +130,7 @@ class TrussHandle:
         """[Deprected] Please use build_serving_docker_image."""
         return self.build_serving_docker_image(*args, **kwargs)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def build_serving_docker_image(
         self,
         build_dir: Optional[Path] = None,
@@ -147,7 +147,7 @@ class TrussHandle:
         self._store_signature()
         return image
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def build_training_docker_image(
         self, build_dir: Optional[Path] = None, tag: Optional[str] = None
     ):
@@ -162,7 +162,7 @@ class TrussHandle:
         """[Deprecated] Do not use."""
         return _docker_image_from_labels(labels)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def docker_run(
         self,
         build_dir: Optional[Path] = None,
@@ -263,7 +263,7 @@ class TrussHandle:
         model = LoadModelLocal.run(self._truss_dir)
         return _prediction_flow(model, request)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def docker_predict(
         self,
         request: Dict,
@@ -313,7 +313,7 @@ class TrussHandle:
 
         return resp.json()
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def docker_train(
         self,
         variables: Optional[dict] = None,
@@ -377,7 +377,7 @@ class TrussHandle:
     def local_train(self, variables: Optional[dict] = None):
         LocalTrainer.run(self._truss_dir)(variables)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def docker_build_setup(self, build_dir: Optional[Path] = None):
         """
         Set up a directory to build docker image from.
@@ -399,7 +399,7 @@ class TrussHandle:
         server_loader = LocalServerLoader(self._truss_dir, image_builder, port=port)
         server_loader.watch(build_dir, work_dir)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def training_docker_build_setup(self, build_dir: Optional[Path] = None):
         """
         Set up a directory to build training docker image from.
@@ -625,7 +625,7 @@ class TrussHandle:
             examples[index] = Example(example_name, example_input)
         self.update_examples(examples)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def get_all_docker_images(self):
         """Returns all docker images for this truss.
 
@@ -633,12 +633,12 @@ class TrussHandle:
         """
         return get_images({TRUSS_DIR: str(self._truss_dir)})
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def get_docker_containers_from_labels(self, *args, **kwargs):
         """[Deprecated] Please use get_serving_docker_containers_from_labels."""
         return self.get_serving_docker_containers_from_labels(*args, **kwargs)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def get_serving_docker_containers_from_labels(
         self,
         all: bool = False,
@@ -669,7 +669,7 @@ class TrussHandle:
         if containers is not None and len(containers) > 0:
             return containers[0]
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def kill_container(self):
         """Kill container
 
@@ -677,12 +677,12 @@ class TrussHandle:
         """
         kill_containers({TRUSS_DIR: self._truss_dir})
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def container_logs(self, *args, **kwargs):
         """[Deprecate] Use serving_container_logs."""
         return self.serving_container_logs(*args, **kwargs)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def serving_container_logs(self, follow=True, stream=True):
         """Get container logs for truss."""
         containers = self.get_serving_docker_containers_from_labels(all=True)
@@ -725,7 +725,7 @@ class TrussHandle:
 
         self._update_config(define_base_image_fn)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def patch_container(self, patch_request: Dict):
         """Patch changes onto the container running this Truss.
 
@@ -751,7 +751,7 @@ class TrussHandle:
         """[Deprecated] Use truss_hash_on_serving_container."""
         return self.truss_hash_on_serving_container()
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def truss_hash_on_serving_container(self) -> Optional[str]:
         """Get content hash of truss running on container."""
         if not self.spec.live_reload:
@@ -804,7 +804,7 @@ class TrussHandle:
     def is_control_truss(self):
         return self._spec.live_reload
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def get_urls_from_truss(self):
         urls = []
         containers = self.get_serving_docker_containers_from_labels()
@@ -831,7 +831,7 @@ class TrussHandle:
 
         self._update_config(enable_live_reload_fn)
 
-    @proxy_to_shadow
+    @proxy_to_shadow_if_scattered
     def calc_patch(self, prev_truss_hash: str) -> Optional[PatchDetails]:
         """Calculates patch of current truss from previous.
 
@@ -858,16 +858,19 @@ class TrussHandle:
         )
 
     def gather(self) -> Path:
-        """Convert a Truss with external dependencies into one without and filter out
-        any files that are defined in the .truss_ignore file.
+        """Convert a Truss with external dependencies into one without.
 
         Any external packages are copied under packages folder to form a Truss,
-        where no parts of the Truss are outside the Truss folder. If a Truss has
-        no external dependencies, then the same Truss is copied to the gathered Truss.
-        We filter out any files that are defined in the .truss_ignore file and return a
-        handle to the gathered Truss. These gathered trusses are cached and resused.
+        where no parts of the Truss are outside the Truss folder. If the Truss
+        doesn't have any external dependencies then this returns the handle to
+        itself. Otherwise, a new truss is created with external dependencies
+        gatherer and a handle to that truss is returned. These gathered trusses
+        are caches and resused.
         """
         from truss.truss_gatherer import gather
+
+        if not self.is_scattered():
+            return self._truss_dir
 
         return gather(self._truss_dir)
